[SYSTEMDS-335] Weighted eviction policy This patch contains a new eviction policy for lineage cache. A min-heap based priority queue over a function of computation time and size is maintained to define the order of evictions.The idea is to evict large matrices, which take little time to recompute. This weighted scheme significantly reduces the number of evictions (including disk spilling). This patch also refactors the LineageCache class to hide the eviction policy related maintenance.
